move assert_msg_in_logs to conftest.py so it can be imported by different unit tests
