+++ tests/conftest.py
@@ -214,3 +214,23 @@ def not_raises(exception):
         yield
     except exception:
         raise pytest.fail(f"DID RAISE EXCEPTION: {exception}")
+
+
+def assert_msg_in_logs(msg: str, msg_logs) -> bool:
+    """Helper method to test if given message is in a list of logged messages.
+
+    Args:
+        msg (str): message to check for
+        msg_logs (List[LogRecord]): log of messages
+
+    Returns:
+        bool: True if msg in msgs, otherwise False
+    """
+    print(f"msg_to_check={msg}")
+    print(f"type={type(msg_logs)}")
+    for i, log in enumerate(msg_logs):
+        the_msg = log.getMessage()
+        print(f"{i} {the_msg}")
+        if msg == the_msg:
+            return True
+    return False
